requirements.txt
# Backend Dependencies

flask==2.3.3
flask-cors==4.0.0
flask-restx==1.3.0
google-cloud-firestore==2.13.1
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
marshmallow==3.20.1
